Feature/#11 deployment average (#75)
* #11: Add average SQL query and split count selection from main queries.
Combine count and average services using a map to select the queries.
Fix monthly query to match date format of quarterly, so that ordering results is consistent.

* #11: Add tests using the mock average map and simplify mock construction.

* #11: Fix name of instance param in dfService serve request.

+++ devlake-go/api/models/response.go
@@ -1,8 +1,8 @@
 package models
 
 type DataPoint struct {
-	Key   string `json:"key" db:"data_key"`
-	Value int    `json:"value" db:"data_value"`
+	Key   string  `json:"key" db:"data_key"`
+	Value float32 `json:"value" db:"data_value"`
 }
 
 type Response struct {
